requirements.txt
openai-whisper
sounddevice
numpy
torch
gradio>=4.0.0

